+++ apollo-biz/src/main/java/com/ctrip/framework/apollo/biz/service/ServerConfigService.java
@@ -18,6 +18,7 @@ package com.ctrip.framework.apollo.biz.service;
 
 import com.ctrip.framework.apollo.biz.entity.ServerConfig;
 import com.ctrip.framework.apollo.biz.repository.ServerConfigRepository;
+import com.ctrip.framework.apollo.common.exception.NotFoundException;
 import com.google.common.collect.Lists;
 import java.util.List;
 import java.util.Objects;
@@ -44,14 +45,15 @@ public class ServerConfigService {
 
   @Transactional
   public ServerConfig createOrUpdateConfig(ServerConfig serverConfig) {
+    if (Objects.isNull(serverConfig.getCluster())) {
+      serverConfig.setCluster("default");
+    }
 
-    ServerConfig storedConfig = serverConfigRepository.findByKey(serverConfig.getKey());
+    ServerConfig storedConfig = serverConfigRepository.findTopByKeyAndCluster(serverConfig.getKey(),
+        serverConfig.getCluster());
 
     if (Objects.isNull(storedConfig)) {// create
       serverConfig.setId(0L);// 为空，设置ID 为0，jpa执行新增操作
-      if (Objects.isNull(serverConfig.getCluster())) {
-        serverConfig.setCluster("default");
-      }
       return serverConfigRepository.save(serverConfig);
     }
 
@@ -61,6 +63,19 @@ public class ServerConfigService {
     storedConfig.setValue(serverConfig.getValue());
 
     return serverConfigRepository.save(storedConfig);
+  }
+
+  @Transactional
+  public void deleteConfig(String key, String cluster, String operator) {
+    ServerConfig storedConfig = serverConfigRepository.findTopByKeyAndCluster(key, cluster);
+
+    if (Objects.isNull(storedConfig)) {
+      throw new NotFoundException("server config not found for key:%s, cluster:%s", key, cluster);
+    }
+
+    storedConfig.setDeleted(true);
+    storedConfig.setDataChangeLastModifiedBy(operator);
+    serverConfigRepository.save(storedConfig);
   }
 
 }