+++ apollo-portal/src/main/resources/static/scripts/controller/ServerConfigController.js
@@ -25,14 +25,23 @@ function ServerConfigController($scope, $window, $translate, toastr, AppUtil, Se
     $scope.portalDBFilterConfigs = [];
     $scope.configDBConfigs = [];
     $scope.configDBFilterConfigs = [];
+    $scope.configDBClusters = ['default'];
+    $scope.showCustomConfigDBCluster = false;
+    $scope.selectedConfigDBCluster = 'default';
     $scope.envs = [];
     $scope.selectedEnv = '';
     $scope.displayModule = 'home';
     $scope.portalDBConfigSearchKey = '';
     $scope.configDBConfigSearchKey = '';
+    $scope.toDeletePortalDBConfig = {};
+    $scope.toDeleteConfigDBConfig = {};
     $scope.configEdit = configEdit;
     $scope.createPortalDBConfig = createPortalDBConfig;
     $scope.createConfigDBConfig = createConfigDBConfig;
+    $scope.deletePortalDBConfig = deletePortalDBConfig;
+    $scope.deleteConfigDBConfig = deleteConfigDBConfig;
+    $scope.confirmDeletePortalDBConfig = confirmDeletePortalDBConfig;
+    $scope.confirmDeleteConfigDBConfig = confirmDeleteConfigDBConfig;
     $scope.gobackPortalDBTabs = gobackPortalDBTabs;
     $scope.gobackConfigDBTabs = gobackConfigDBTabs;
     $scope.portalDBConfigFilter = portalDBConfigFilter;
@@ -40,6 +49,8 @@ function ServerConfigController($scope, $window, $translate, toastr, AppUtil, Se
     $scope.resetPortalDBConfigSearchKey = resetPortalDBConfigSearchKey;
     $scope.resetConfigDBConfigSearchKey = resetConfigDBConfigSearchKey;
     $scope.switchConfigDBEnvs = switchConfigDBEnvs;
+    $scope.useCustomConfigDBCluster = useCustomConfigDBCluster;
+    $scope.useConfigDBClusterSelect = useConfigDBClusterSelect;
 
     $scope.allowSwitchingTabs = true;
 
@@ -88,25 +99,67 @@ function ServerConfigController($scope, $window, $translate, toastr, AppUtil, Se
                 $scope.configDBConfigs.push(user);
                 $scope.configDBFilterConfigs.push(user);
             });
+            refreshConfigDBClusters(result);
         },function (result) {
             $scope.configDBConfigs = [];
             $scope.configDBFilterConfigs = [];
+            refreshConfigDBClusters([]);
             toastr.error(AppUtil.errorMsg(result), $translate.instant('Config.SystemError'));
         });
     }
 
-    function configEdit (displayModule,config) {
+    function refreshConfigDBClusters(configs) {
+        var clusters = ['default'];
+        configs.forEach(function (config) {
+            if (config.cluster && clusters.indexOf(config.cluster) < 0) {
+                clusters.push(config.cluster);
+            }
+        });
+        $scope.configDBClusters = clusters;
+    }
+
+    function configEdit (displayModule, config, isConfigDB) {
         $scope.displayModule = displayModule;
         $scope.allowSwitchingTabs = false;
+        $scope.showCustomConfigDBCluster = false;
 
         $scope.serverConfig = {};
         if (config != null) {
+            var cluster = config.cluster || 'default';
             $scope.serverConfig = {
                 key: config.key,
+                cluster: cluster,
                 value: config.value,
                 comment: config.comment
             };
+            $scope.selectedConfigDBCluster = cluster;
+        } else if (isConfigDB) {
+            $scope.serverConfig.cluster = 'default';
+            $scope.selectedConfigDBCluster = 'default';
         }
+    }
+
+    function useCustomConfigDBCluster() {
+        $scope.showCustomConfigDBCluster = true;
+        if ($scope.serverConfig.cluster === 'default') {
+            $scope.serverConfig.cluster = '';
+        }
+    }
+
+    function useConfigDBClusterSelect() {
+        $scope.showCustomConfigDBCluster = false;
+        var selectedCluster = $scope.serverConfig.cluster || $scope.selectedConfigDBCluster || 'default';
+
+        if ($scope.configDBClusters.indexOf(selectedCluster) < 0) {
+            if ($scope.configDBClusters.indexOf($scope.selectedConfigDBCluster) >= 0) {
+                selectedCluster = $scope.selectedConfigDBCluster;
+            } else {
+                selectedCluster = 'default';
+            }
+        }
+
+        $scope.serverConfig.cluster = selectedCluster;
+        $scope.selectedConfigDBCluster = selectedCluster;
     }
 
     function switchConfigDBEnvs(env) {
@@ -126,6 +179,9 @@ function ServerConfigController($scope, $window, $translate, toastr, AppUtil, Se
     }
 
     function createConfigDBConfig() {
+        if (!$scope.showCustomConfigDBCluster) {
+            $scope.serverConfig.cluster = $scope.selectedConfigDBCluster || 'default';
+        }
         ServerConfigService.createConfigDBConfig($scope.selectedEnv, $scope.serverConfig).then(function (result) {
             toastr.success($translate.instant('ServiceConfig.Saved'));
             getConfigDBConfig();
@@ -133,6 +189,46 @@ function ServerConfigController($scope, $window, $translate, toastr, AppUtil, Se
             $scope.allowSwitchingTabs = true;
         }, function (result) {
             toastr.error(AppUtil.errorMsg(result), $translate.instant('ServiceConfig.SaveFailed'));
+        });
+    }
+
+    function deletePortalDBConfig(config) {
+        $scope.toDeletePortalDBConfig = {
+            key: config.key
+        };
+        $('#deletePortalDBConfirmDialog').modal('show');
+    }
+
+    function confirmDeletePortalDBConfig() {
+        ServerConfigService.deletePortalDBConfig($scope.toDeletePortalDBConfig.key).then(function () {
+            toastr.success($translate.instant('ServiceConfig.Deleted'));
+            getPortalDBConfig();
+            $scope.toDeletePortalDBConfig = {};
+        }, function (result) {
+            toastr.error(AppUtil.errorMsg(result), $translate.instant('ServiceConfig.DeleteFailed'));
+        });
+    }
+
+    function deleteConfigDBConfig(config) {
+        $scope.toDeleteConfigDBConfig = {
+            env: $scope.selectedEnv,
+            key: config.key,
+            cluster: config.cluster || 'default'
+        };
+        $('#deleteConfigDBConfirmDialog').modal('show');
+    }
+
+    function confirmDeleteConfigDBConfig() {
+        ServerConfigService.deleteConfigDBConfig(
+            $scope.toDeleteConfigDBConfig.env,
+            $scope.toDeleteConfigDBConfig.key,
+            $scope.toDeleteConfigDBConfig.cluster
+        ).then(function () {
+            toastr.success($translate.instant('ServiceConfig.Deleted'));
+            getConfigDBConfig();
+            $scope.toDeleteConfigDBConfig = {};
+        }, function (result) {
+            toastr.error(AppUtil.errorMsg(result), $translate.instant('ServiceConfig.DeleteFailed'));
         });
     }